Move dioxus-specifc code into dioxus-tui crate and rename core crate to Rink
