+++ packages/dioxus-tui/examples/all_events.rs
