+++ packages/dioxus-tui/examples/quadrants.rs
