+++ packages/dioxus-tui/examples/stress.rs
