+++ packages/dioxus-tui/src/prelude/mod.rs
@@ -1,0 +1,2 @@
+pub use crate::widgets::*;
+pub use rink::Config;
