+++ packages/rink/.vscode/spellright.dict
@@ -1,0 +1,2 @@
+esque
+Tui
