+++ packages/rink/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "dioxus-tui"
+name = "rink"
 version = "0.2.2"
 authors = ["Jonathan Kelley, @dementhos"]
 edition = "2021"
@@ -13,12 +13,9 @@ license = "MIT/Apache-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-dioxus = { path = "../dioxus", version = "^0.3.0", optional = true }
-dioxus-core = { path = "../core", version = "^0.3.0", features = ["serialize"], optional = true }
 dioxus-html = { path = "../html", version = "^0.3.0" }
 dioxus-native-core = { path = "../native-core", version = "^0.2.0" }
 dioxus-native-core-macro = { path = "../native-core-macro", version = "^0.3.0" }
-dioxus-hot-reload = { path = "../hot-reload", optional = true }
 
 tui = "0.17.0"
 crossterm = "0.23.0"
@@ -34,15 +31,9 @@ shipyard = { version = "0.6.2", features = ["proc", "std"], default-features = f
 once_cell = "1.17.1"
 
 [dev-dependencies]
-dioxus = { path = "../dioxus" }
 tokio = { version = "1" }
 criterion = "0.3.5"
 
-[[bench]]
-name = "update"
-harness = false
-
 [features]
-default = ["hot-reload", "dioxus-bindings"]
-hot-reload = ["dioxus-hot-reload"]
-dioxus-bindings = ["dioxus", "dioxus-core", "dioxus-native-core/dioxus"]
+default = []
+parallel = ["shipyard/parallel"]
