+++ packages/rink/src/focus.rs
@@ -19,8 +19,9 @@ use dioxus_native_core::node_ref::NodeView;
 #[derive(Component)]
 pub struct Focused(pub bool);
 
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug, Default)]
 pub(crate) enum FocusLevel {
+    #[default]
     Unfocusable,
     Focusable,
     Ordered(std::num::NonZeroU16),
@@ -55,12 +56,6 @@ impl PartialOrd for FocusLevel {
 impl Ord for FocusLevel {
     fn cmp(&self, other: &Self) -> std::cmp::Ordering {
         self.partial_cmp(other).unwrap()
-    }
-}
-
-impl Default for FocusLevel {
-    fn default() -> Self {
-        FocusLevel::Unfocusable
     }
 }
